Add macos-14 to test_python matrix
macos-14 is aarch64 with Apple silicon.  This lets us test the aarch64
wheel in CI

may try a follow up to see if the builds are any faster on macos-14.
Will probably want to double check we still have a good value for
`CMAKE_OSX_DEPLOYMENT_TARGET`/`MACOSX_DEPLOYMENT_TARGET` if upgrading
the builders

Signed-off-by: Tom Jakubowski <tom@prospective.dev>