revert me revert me comment out exception test
OMM, using the wheel built in CI, this test and only this test hangs.
the others pass

still not getting test output in CI after the "collected" line -- that
could be because all the tests are hanging, or it could be an issue in
flushing pytest's output or something.  curious to see if disabling just
this test will turn the emscripten wheel tests green

deliberately not signing so it doesn't get merged